# backend-api/docs/swagger.yaml
swagger: '2.0'
info:
    title: Ostukorvi optimeerija
    version: 1.0.0
    description: Sisesta tooteid ja hinda
paths:
    /products:
        get:
            tags:
                - Products
            responses:
                '200':
                    description: Returns list of Products
                    schema:
                        type: array
                        items:
                            type: string
                    examples:
                        application/json:
                            - Piim
                            - Leib
            summary: Get all the product in API
            description: Return all Products that are registered in API
        post:
            tags:
                - Products
            parameters:
                -
                    name: body
                    description: |
                        The Product that is to be created.
                    schema:
                        $ref: '#/definitions/new-Product'
                    in: body
                    required: true
            responses:
                '201':
                    description: Send 201 when Product is successfully added to the database.
                    schema:
                        $ref: '#/definitions/new-Product'
                '400':
                    description: 'Missing parameter, please review your request data.'
            summary: Create a new product into the app
            description: |
                Creates a new Product instance with its required details for further usage.
    '/products/{ProductID}':
        get:
            tags:
                - Products
            responses:
                '200':
                    description: Returns all details of the given Product
                    schema:
                        type: array
                        items:
                            type: string
                    examples:
                        application/json:
                            ProductID: 6b52cae6-cec3-11f0-b43f-901b0eec5991
                            Name: TestProduct
                            Category: TestCategory
                            Price: 6
                            Description: ''
                            createdAt: '2025-12-01T16:37:17.000Z'
                            updatedAt: '2025-12-01T16:37:17.000Z'
                '404':
                    description: Returns error when Product not found
                    schema:
                        $ref: '#/definitions/error'
            operationId: ProductID
            summary: Get full details of a Product
            description: Will return all details of a given Product based in ProductID given
        put:
            tags:
                - Products
            parameters:
                -
                    name: body
                    description: Contains new data to be changed in the specified Product
                    schema:
                        $ref: '#/definitions/new-Product'
                    in: body
                    required: true
            responses:
                '201':
                    description: Return 201 upon successful modification of Product
                    schema:
                        $ref: '#/definitions/new-Product'
                    examples:
                        application/json:
                            -
                                Name: Spagheti
                                Price: 7.65
                                Category: Pasta
                                Description: Good pasta
                '400':
                    description: Returns 400 when one or multiple required parameters are missing.
                    schema:
                        $ref: '#/definitions/error'
                    examples:
                        application/json:
                            -
                                error: 'Can''t change info, because some parameter is missing.'
                '404':
                    description: 'Returns 404 when Product that is to be modified, is not found by the given ProductID'
                    schema:
                        $ref: '#/definitions/error'
                    examples:
                        application/json:
                            -
                                error: Product not found
            operationId: ProductID
            summary: Modify details of Product
        delete:
            tags:
                - Products
            parameters:
                -
                    name: body
                    description: ID of the Product to be removed from the app.
                    schema:
                        $ref: '#/definitions/delete-Product'
                    in: body
                    required: true
            responses:
                '204':
                    description: Returns "No Content" when removal of Product is successful.
                    schema:
                        $ref: '#/definitions/error'
                    examples:
                        application/json:
                            -
                                error: No Content
                '404':
                    description: |
                        Return error when Product with specified ID does not exist.
                    schema:
                        $ref: '#/definitions/error'
                    examples:
                        application/json:
                            -
                                error: ID Not Found
            operationId: ProductID
            summary: Delete a Product from the app
            description: |
                Deletes an entry from the app.
        parameters:
            -
                name: ProductID
                description: ID of Product
                in: path
                required: true
                type: string
    /profiles:
        get:
            tags:
                - Profiles
            responses:
                '200':
                    description: Returns list of profiles
                    schema:
                        type: array
                        items:
                            type: string
                    examples:
                        application/json: |4-

                                [
                                    "test@gm.com",
                                    "suva@ke.com"
                                ]
            summary: Get all profiles in api
            description: Rerurn all profiles that are registred in api
        post:
            tags:
                - Profiles
            parameters:
                -
                    name: body
                    description: Data of new profile to be added
                    schema:
                        $ref: '#/definitions/new-Profile'
                    in: body
                    required: true
            responses:
                '201':
                    description: Send 201 when Profile is successfully added to the database.
                    schema:
                        $ref: '#/definitions/new-Profile'
                '400':
                    description: 'Missing parameters, please review your request data.'
                    schema:
                        $ref: '#/definitions/error'
            summary: Create a new profile into the app
            description: Creates a new Profile instance with its required details for further usage.
    /baskets:
        get:
            tags:
                - Baskets
            responses:
                '200':
                    description: Returns a list of all the Baskets.
                    schema:
                        type: array
                        items:
                            type: string
                    examples:
                        application/json:
                            - Every day products list
                            - My wishlist
            summary: Get all Baskets in the app
            description: Return all the Baskets registered in you API.
        post:
            tags:
                - Baskets
            parameters:
                -
                    name: body
                    description: Basket that is to be created.
                    schema:
                        $ref: '#/definitions/new-Basket'
                    in: body
                    required: true
            responses:
                '201':
                    description: Send 201 when Basket is successfully created.
                    schema:
                        $ref: '#/definitions/new-Basket'
                '400':
                    description: 'Missing parameter, please review your request data.'
            summary: Create a new Basket into the app
            description: Creata a new Basket with all the required fields.
    '/baskets/{BasketID}':
        get:
            tags:
                - Baskets
            responses:
                '200':
                    description: Returns all the details of the given Basket
                    schema:
                        type: array
                        items: {}
                    examples:
                        application/json: |-
                            {
                                "BasketID": "6b52cah6-cec3-11f0-b43f-901b0eec5991",
                                "Name": "TestBasket",
                                "Description": "TestBasket description",
                                "TotalPrice": 49.98
                                "createdAt": "2025-12-01T16:37:17.000Z",
                                "updatedAt": "2025-12-01T16:37:17.000Z"
                            }
                '404':
                    description: Returns error when Basket isn't found
                    schema:
                        $ref: '#/definitions/error'
            operationId: BasketID
            summary: Get full details of the Basket
            description: Will return all the details of given Basket when BasketID is given
        put:
            tags:
                - Baskets
            parameters:
                -
                    name: body
                    description: Contains new data for selected Basket.
                    schema:
                        $ref: '#/definitions/new-Basket'
                    in: body
                    required: true
            responses:
                '201':
                    description: Returns 201 after Basket has been successfully modified with new data.
                    schema:
                        $ref: '#/definitions/new-Basket'
                    examples:
                        application/json: |-
                            [
                                {
                                    "Name": "Modified basket name",
                                    "Description": "Very good basket",
                                }
                            ]
                '400':
                    description: Returns 400 when one or more parameters are missing.
                    schema:
                        $ref: '#/definitions/error'
                '404':
                    description: 'Returns 404, when Basket that is to be modified, isn''t found by given BasketID'
                    schema:
                        $ref: '#/definitions/error'
            operationId: BasketID
            summary: Modify selected Basket
        delete:
            tags:
                - Baskets
            responses:
                '204':
                    description: Returns 204 when removal of the selected Basket is successful.
                    schema:
                        $ref: '#/definitions/error'
                    examples:
                        application/json:
                            -
                                error: No Content
                '404':
                    description: Returns 404 when Basket isn't found by the given BasketID.
                    schema:
                        $ref: '#/definitions/error'
                    examples:
                        application/json:
                            -
                                error: ID not found
            operationId: BasketID
            summary: Delete selected Basket from the app
        parameters:
            -
                name: BasketID
                description: Id of the Basket
                in: path
                required: true
                type: string
    '/profiles/{ProfileID}':
        get:
            tags:
                - Profiles
            responses:
                '200':
                    description: Returns all details of the given Profile
                    schema:
                        type: array
                        items:
                            type: string
                    examples:
                        application/json:
                            UserID: 6b52cae6-cec3-11f0-b43f-901b0eec5991
                            Email: asd@mail.ee
                            IsAdmin: 'false'
                            createdAt: '2015-12-01T16:37:17.000Z'
                            updatedAt: '2015-12-01T16:37:17.000Z'
                '404':
                    description: Returns error when Profile not found
                    schema:
                        $ref: '#/definitions/error'
            operationId: ProfileID
            summary: Get full details of a Profile
            description: Will return all details of a given Profile based in ProfileId given
        put:
            tags:
                - Profiles
            parameters:
                -
                    name: body
                    description: Contains new data to be changed in the specified Profile.
                    schema:
                        $ref: '#/definitions/new-Profile'
                    in: body
                    required: true
            responses:
                '201':
                    description: Return 201 upon successful modification of Profile
                    schema:
                        $ref: '#/definitions/new-Product'
                    examples:
                        application/json:
                            -
                                Email: uxui@mail.ee
                                Password: testpassword
                '400':
                    description: Returns 200 when one or multiple required parameters are missing.
                    schema:
                        $ref: '#/definitions/error'
                    examples:
                        application/json:
                            -
                                error: 'Can`t change info, because some parameter is missing'
                '404':
                    description: 'Returns 404 when Profile that is to be mofified, is not found by the given ProfileID'
                    schema:
                        $ref: '#/definitions/error'
                    examples:
                        application/json:
                            -
                                error: Profile not found
            operationId: ProfileID
            summary: Modify details of Profile
        delete:
            tags:
                - Profiles
            responses:
                '204':
                    description: Returns "No Content" when removal of Profile is succesful
                    schema:
                        $ref: '#/definitions/error'
                    examples:
                        application/json:
                            -
                                error: No Content
                '404':
                    description: Return error when Profile with specified ID does not exist
                    schema:
                        $ref: '#/definitions/error'
                    examples:
                        application/json:
                            -
                                error: ID Not Found
            operationId: ProfileID
            summary: Delete a Profile from the app
            description: Deletes an entry from the app.
        parameters:
            -
                name: ProfileID
                in: path
                required: true
definitions:
    error:
        title: Root Type for error
        description: ''
        type: object
        properties:
            error:
                type: string
        example:
            error: Error message
    new-Product:
        title: Root Type for new-Product
        description: ''
        required:
            - Category
            - Description
            - Name
            - Price
        type: object
        properties:
            Name:
                type: string
            Price:
                format: double
                type: number
            Category:
                type: string
            Description:
                type: string
        example:
            Name: NewProductName
            Price: 9.99
            Category: NewProductCategory
            Description: NewProductDescription
    delete-Product:
        title: Root Type for delete-Product
        description: ''
        required:
            - ProductID
        type: object
        properties:
            ProductID:
                type: string
        example:
            ProductID: 6b52cae6-cec3-11f0-b43f-901b0eec5991
    new-Profile:
        title: Root Type for new-Profile
        description: This data is currently a complex object.
        required:
            - Password
            - Email
        type: object
        properties:
            Email:
                type: string
            Password:
                type: string
            IsAdmin:
                description: ''
                type: boolean
        example:
            Email: xxx@sample.ee
            PasswordHASH: xxxxx
            IsAdmin: false
    new-Basket:
        title: Root Type for new-Basket
        description: ''
        required:
            - Name
            - Description
        type: object
        properties:
            Name:
                type: string
            Description:
                type: string
            TotalPrice:
                format: double
                type: number
        example:
            Name: Example basket name
            Description: This is basket for my example
    delete-Basket:
        title: Root Type for delete-Basket
        description: ''
        required:
            - BasketID
        type: object
        properties:
            BasketID:
                type: string
        example:
            BasketID: 19afec7-c2f3-72c8-b83f-bebcb35fc358
    delete-Profile:
        title: Root Type for delete-Profile
        description: ''
        required:
            - ProfileID
        type: object
        properties:
            ProfileID:
                type: string
        example:
            ProfileID: 6b52cae6-cec3-11f0-b43f-901b0eec5991